+++ CHANGELOG.md
@@ -6,9 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added 
 -`isAvailable` flag to the productDataLayer
+
 ## [3.0.1] - 2021-07-29
 ### Fixed
 - Add originalLocation variable to dataLayer