+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.0.3] - 2021-10-29
+
 ### Fixed:
 - The variable name `OriginalLocation` to `originalLocation`.